Move GcLogger to its own file; fixes #4194
